FIX | timer repeat from uint8_t type to int16_t

+++ core/dev_event_timer.h
@@ -10,7 +10,7 @@ typedef int  (*timer_handler_t)(void *ev_data, void *self_data);
 
 typedef struct _dev_timer_ev_t 
 {
-    uint8_t repeat;        
+    int16_t repeat;        
     double timeout;
     struct timespec ts;
     timer_handler_t cb;
